+++ mora/HW.py
@@ -15,6 +15,7 @@ class DLA(object):
         self.dataflow = dataflow
         self.home_path = home_path
         self.dla_dicts = self.set_dla()
+        self.DSE_indicator = 0
 
     def set_dla(self):
         dla_dicts = {}
@@ -26,8 +27,12 @@ class DLA(object):
         dla_dicts['dataflow'] = self.dataflow
         return dla_dicts
 
+    def set_dse_param(self, pes, noc_bw, indicator):
+        self.dla_dicts['pes'] = pes
+        self.dla_dicts['noc_bw'] = noc_bw
+        self.DSE_indicator = indicator
+
     def invoke_maestro(self, model):
-        output_csv_path = os.path.abspath(os.path.join(self.home_path, 'output/' + model + '/' + model + '-dla.csv'))
         mapping_path = os.path.abspath(os.path.join(self.home_path, 'model/' + model + '/' + model + '-dla_' + self.dataflow + '.m'))
         maestro_result_csv_path = os.path.abspath(os.path.join(self.home_path, 'output/' + model + '/' + model + '-dla_' + self.dataflow + '.csv'))
         maestro_path = os.path.abspath(os.path.join(self.home_path, 'maestro'))
@@ -54,27 +59,39 @@ class DLA(object):
         stdout, stderr = process.communicate()
         process.wait()
 
+    def export(self, model, on_DLA_layer_index=None):
+        output_csv_path = os.path.abspath(os.path.join(self.home_path, 'output/' + model + '/' + model + '-dla.csv'))
+        maestro_result_csv_path = os.path.abspath(os.path.join(self.home_path, 'output/' + model + '/' + model + '-dla_' + self.dataflow + '.csv'))
         # maestro original maestro_result csv pisition: maestro/.
         output_csv_dicts = {}
         try:
             maestro_result_df = pd.read_csv(maestro_result_csv_path)
             layers = maestro_result_df.shape[0]
-            runtime_nd = maestro_result_df[' Runtime (Cycles)'].to_numpy().reshape(-1, 1)
-            energy_nd = maestro_result_df[' Activity count-based Energy (nJ)'].to_numpy().reshape(-1, 1)
+            if self.DSE_indicator != 0:
+                on_DLA_layer_index = range(layers)
+            elif on_DLA_layer_index is None:
+                raise AttributeError
+                sys.exit(1)
+            runtime_nd = maestro_result_df[' Runtime (Cycles)'].to_numpy().reshape(-1, 1)[on_DLA_layer_index]
+            energy_nd = maestro_result_df[' Activity count-based Energy (nJ)'].to_numpy().reshape(-1, 1)[on_DLA_layer_index]
             area = maestro_result_df.at[0, ' Area']
-            power_nd = maestro_result_df[' Power'].to_numpy().reshape(-1, 1)
-            l2_size_nd = maestro_result_df["  L2 SRAM Size Req (Bytes)"].to_numpy().reshape(-1, 1)
+            power_nd = maestro_result_df[' Power'].to_numpy().reshape(-1, 1)[on_DLA_layer_index]
+            l2_size_nd = maestro_result_df["  L2 SRAM Size Req (Bytes)"].to_numpy().reshape(-1, 1)[on_DLA_layer_index]
             if np.median(l2_size_nd) > self.dla_dicts['glb_size']:
                 print('maestro glb size exeed.')
                 return
+            output_csv_dicts['DSE index'] = self.DSE_indicator
             output_csv_dicts['layers'] = layers
             output_csv_dicts['latency'] = runtime_nd.sum()
             output_csv_dicts['energy'] = energy_nd.sum()
             output_csv_dicts['area'] = area
             output_csv_dicts['power'] = power_nd.mean()
+            output_csv_dicts['restraint'] = 'pass'
             csv = pd.DataFrame(output_csv_dicts, index=[0])
-            csv.to_csv(output_csv_path)
-            # print('maestro invoked.')
+            if os.path.exists(output_csv_path):
+                csv.to_csv(output_csv_path, mode='a', header=False)
+            else:
+                csv.to_csv(output_csv_path)
         except FileNotFoundError:
             print('maestro invoke fatal.')
         return
@@ -88,6 +105,7 @@ class RRAM(object):
         self.mem_capacity = self.get_memory_capacity()
         self.home_path = home_path
         self.rram_config_path = os.path.abspath(os.path.join(home_path, "rram_config.ini"))
+        self.DSE_indicator = 0
 
     def set_rram(self):
         rram_dicts = {}
@@ -97,8 +115,15 @@ class RRAM(object):
         rram_dicts['tile_bw'] = self.hw_param_dicts['rram_tile_bw']
         return rram_dicts
 
+    def set_dse_param(self, row, col, tile_bw, indicator):
+        self.rram_dicts['tile_row'] = row
+        self.rram_dicts['tile_col'] = col
+        self.rram_dicts['tile_bw'] = tile_bw
+        self.mem_capacity = self.get_memory_capacity()
+        self.DSE_indicator = indicator
+
     def get_memory_capacity(self):
-        return self.rram_dicts['tile_row'] * self.rram_dicts['tile_col'] * 16 * 8 * 2 * 128**2  # bit
+        return self.rram_dicts['tile_row'] * self.rram_dicts['tile_col'] * 16 * 8 * 2 * 128**2 * 2  # bit
 
     def invoke_MNSIM(self, model):
         output_csv_path = os.path.abspath(os.path.join(self.home_path, 'output/' + model + '/' + model + '-rram.csv'))
@@ -106,7 +131,7 @@ class RRAM(object):
         #    print("rram outfile conflict.")
         #    raise AttributeError
         print("[MNSIM] Result", self.rram_dicts)
-        import_module("MNSIM_main").main(model, [self.rram_dicts['tile_row'], self.rram_dicts['tile_col'], self.rram_dicts['tile_bw']])
+        import_module("MNSIM_main").main(model, [self.rram_dicts['tile_row'], self.rram_dicts['tile_col']], self.rram_dicts['tile_bw'], self.DSE_indicator)
         '''
         command = [
             "python", "../MNSIM.py", "--model {}".format(model), "--tile_size {} {}".format(self.rram_dicts['tile_row'], self.rram_dicts['tile_col']),
